feat(vpp): stabilize eager VPP with PD and async scheduling
Implement virtual pipeline parallelism and preserve the validated eager-mode compatibility path.

Add fold-back communication warmup, split-queue liveness handling, bounded dummy futures, deferred async sampling state, Mooncake fold-back layer discovery, and pooled input-batch cleanup.

+++ vllm_ascend/ops/fused_moe/token_dispatcher.py
@@ -621,7 +621,8 @@ class TokenDispatcherWithAll2AllV(MoETokenDispatcher[MoEAllToAllCombineMetadata]
             if num_global_tokens_per_local_expert is None:
                 raise ValueError("num_global_tokens_per_local_expert must be set before operations.")
             global_input_tokens_local_experts_indices = torch.repeat_interleave(
-                self.expert_ids_per_ep_rank, num_global_tokens_per_local_expert.ravel()
+                self.expert_ids_per_ep_rank,
+                num_global_tokens_per_local_expert.ravel().to(torch.int32),
             )
         else:
             torch.npu.synchronize()
